move component to lab
