+++ docs/pages/api-docs/n-progress-bar.json
@@ -26,7 +26,7 @@
   },
   "spread": true,
   "forwardsRefTo": "HTMLSpanElement",
-  "filename": "/packages/material-ui/src/NProgressBar/NProgressBar.js",
+  "filename": "/packages/material-ui-lab/src/NProgressBar/NProgressBar.js",
   "inheritance": null,
   "demos": "<ul><li><a href=\"/components/n-progress/\">NProgress</a></li></ul>",
   "styledComponent": true,
