+++ packages/material-ui-lab/src/NProgressBar/NProgress.d.ts
