+++ packages/material-ui-lab/src/NProgressBar/index.js
