+++ packages/material-ui-lab/src/NProgressBar/nProgressBarClasses.d.ts
