+++ packages/material-ui-lab/src/NProgressBar/nProgressState.js
