+++ packages/material-ui-lab/src/index.js
@@ -62,6 +62,9 @@ export * from './MobileTimePicker';
 export { default as MonthPicker } from './MonthPicker';
 export * from './MonthPicker';
 
+export { default as NProgressBar } from './NProgressBar';
+export * from './NProgressBar';
+
 export { default as Pagination } from './Pagination';
 export * from './Pagination';
